SIMPLE-DOS: Clean up attacking server
