+++ 00-gladiatorsofjs-exercises/03-Backend-Express-Next-TS/1.Server/solutions/6-simpleDOS/A-server/src/api/sample/attack.controller.ts
@@ -1,0 +1,11 @@
+import { Response, Request } from 'express';
+
+import { attackingService } from './attack.service';
+
+export class AttackController {
+  static attack(req: Request, res: Response) {
+    const result = attackingService.attack();
+
+    res.status(200).send(result);
+  }
+}
