+++ 00-gladiatorsofjs-exercises/03-Backend-Express-Next-TS/1.Server/solutions/6-simpleDOS/A-server/src/api/sample/attack.router.ts
@@ -1,0 +1,20 @@
+import { Router } from 'express';
+
+import { IRouter } from '../../types';
+
+import { AttackController } from './attack.controller';
+
+export class AttackingRouter implements IRouter {
+  public path = '/';
+  public router: Router;
+
+  constructor() {
+    this.router = Router();
+
+    this.initializeRoutes();
+  }
+
+  private initializeRoutes() {
+    this.router.post('/attack', AttackController.attack);
+  }
+}
