+++ 00-gladiatorsofjs-exercises/03-Backend-Express-Next-TS/1.Server/solutions/6-simpleDOS/A-server/src/api/sample/attack.service.ts
@@ -1,0 +1,7 @@
+export class AttackingService {
+  attack() {
+    return 'Attacking...';
+  }
+}
+
+export const attackingService = new AttackingService();
