--- 00-gladiatorsofjs-exercises/03-Backend-Express-Next-TS/1.Server/solutions/6-simpleDOS/A-server/src/api/sample/sample.service.ts
@@ -1,18 +1,0 @@
-import { AppError } from "../../helpers/error";
-import { IRepository, SampleRepository } from "../../types";
-
-export class SampleService {
-  sampleRepository: SampleRepository;
-
-  constructor(repository: SampleRepository) {
-    this.sampleRepository = repository;
-  }
-
-  sayHelloFromUser(name: string) {
-    const user = this.sampleRepository.getUser(name);
-
-    if (!user) throw new AppError(400, "User not found");
-
-    return `Hello from ${user}!`;
-  }
-}
