+++ 00-gladiatorsofjs-exercises/03-Backend-Express-Next-TS/1.Server/solutions/6-simpleDOS/B-server/src/api/images/images.service.ts
@@ -1,5 +1,4 @@
 import path from 'path';
-import fs from 'fs';
 
 export class ImagesService {
   getFirstImage() {
